# Optimize allowlist handling